+++ .next/dev/server/chunks/ssr/_next-internal_server_app_stadiums_page_actions_106v2r_.js
@@ -1,0 +1,7 @@
+module.exports = [
+"[project]/.next-internal/server/app/stadiums/page/actions.js [app-rsc] (server actions loader, ecmascript)", ((__turbopack_context__, module, exports) => {
+
+}),
+];
+
+//# sourceMappingURL=_next-internal_server_app_stadiums_page_actions_106v2r_.js.map