+++ .next/dev/server/chunks/ssr/_next-internal_server_app_teams_page_actions_0swn.z_.js
@@ -1,0 +1,7 @@
+module.exports = [
+"[project]/.next-internal/server/app/teams/page/actions.js [app-rsc] (server actions loader, ecmascript)", ((__turbopack_context__, module, exports) => {
+
+}),
+];
+
+//# sourceMappingURL=_next-internal_server_app_teams_page_actions_0swn.z_.js.map