+++ .next/dev/server/server-reference-manifest.js
@@ -1,0 +1,1 @@
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"fXjyJknTWusEj1WtFvjx8z3V79p7gtHp7kdyP1hquaM=\"\n}"