+++ .next/dev/static/chunks/app_layout_tsx_004glpo._.js
@@ -1,0 +1,5 @@
+(globalThis["TURBOPACK_CHUNK_LISTS"] || (globalThis["TURBOPACK_CHUNK_LISTS"] = [])).push({
+    script: typeof document === "object" ? document.currentScript : undefined,
+    chunks: ["static/chunks/[root-of-the-server]__10.w3k_._.css","static/chunks/node_modules_12kdf87._.js","static/chunks/app_layout_Header_tsx_03pzx.y._.js"],
+    source: "dynamic"
+});